backend/requirements.txt
Flask
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.3
Flask-CORS==3.0.10
Werkzeug==2.2.2
pyjwt